+++ Utility/SDL_Utility.h
@@ -7,9 +7,9 @@
 SDL_Surface* createSurface(int w, int h);
 void setColor(SDL_Surface* surf, int x, int y, const Color& c);
 void setColor(SDL_Surface* surf, int x, int y, Uint32 color);
-void setBlendColor(SDL_Surface *surf, Uint16 x, Uint16 y, const Color& c);
+void setBlendColor(SDL_Surface *surf, int x, int y, const Color& c);
 void setBlendColor(SDL_Surface* surf, Uint32* pixel, const Color& c);
-void setBlendColor(SDL_Surface * surf, Uint16 x, Uint16 y, const Color& c, int weight);
+void setBlendColor(SDL_Surface * surf, int x, int y, const Color& c, int weight);
 
 template<class T>
 inline T lerp(const float f, const T a, const T b) {
@@ -17,3 +17,4 @@ inline T lerp(const float f, const T a, const T b) {
 }
 
 #endif
+
